Evo_synth changes
Created a backup in case the changes that have been made are horribly wrong, but otherwise 'fixed' all of the evo_synth errors except the one relating to the ethernet. 
Note: The base address given for the exceptions for the nios ii exception slave is that of the on chip memory.